requirements.txt
Telethon==1.19.0
python-dotenv==0.15.0
backoff
ratelimiter
tinydb==4.3.0